requirements.txt
streamlit==1.33.0
boto3==1.34.59
botocore==1.34.59
openai==1.12.0
python-dotenv==1.0.1
Pillow==10.2.0
numpy==1.26.4